upload monstersvalley for testing
git-svn-id: https://version-control.adelaide.edu.au/svn/a1691807/2017/s2@896 ebee28f9-b968-474b-8da9-e5d0b78a977b

+++ pssd/week04practice/MonstersValley2.hpp
@@ -30,7 +30,7 @@ class MonstersValley2{
             return checkMin(0,0,dread,price);
         }
 
-        int checkMin(int itr,int curD,vector<int> dread,vector<int> price){
+        int checkMin(int itr,long curD,vector<int> dread,vector<int> price){
             if(itr>=dread.size())
                 return 0;
             if(curD<dread[itr])
